+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 
  - Allow multiple data series on asynchronous graph ([#399](https://github.com/kizniche/mycodo/issues/399))
+ - Add Outputs to Asynchronous Graphs ([#399](https://github.com/kizniche/mycodo/issues/399))
+ - Preserve Asynchronous Graph selections after form submissions ([#399](https://github.com/kizniche/mycodo/issues/399))
 
 ### Bugfixes